fix: update the imports to match the file renames
