+++ src/pybert/models/dfe.py
@@ -14,7 +14,7 @@ Copyright (c) 2014 by David Banas; All rights reserved World wide.
 from numpy import array, sign, zeros
 from scipy.signal import iirfilter
 
-from pybert.sim.cdr import CDR
+from pybert.models.cdr import CDR
 
 gNch_taps = 3  # Number of taps used in summing node filter.
